Clone fswreck here from tools for online filecheck testing
The reasons why doing clone are:
1. we need utility to corrupt a file for testing onlike filecheck
feature; 2. fswreck is not exported by tools so that it's not convenient
to use; 3. fswreck is originally designed for wrecking filesystem, which
is not flexible enough for our needs sometime, for example, it will disappoint
us if I want to corrupt an specified inodeA with INODE_BLKNO type, because
the CLI cannot take inode number as parameter and do black-box operation;
4. It's very boring to continuously add new trivals into tools project;
5. We can remove unneeded code, reconstruct interface and add new functions
later, etc.

This patch just copy fswreck code here. The following patches is making it
build and work.

Signed-off-by: Eric Ren <zren@suse.com>